+++ conf.yml
@@ -1,0 +1,7 @@
+hutch: sxr
+
+# Locate happi database
+db: /reg/g/pcds/pyps/apps/hutch-python/device_config/db.json
+
+# Hutch-specific imports
+load: sxr.beamline
