+++ sxr/exceptions.py
