+++ sxr/tests/test_palette.py
