+++ sxr/tests/test_plans.py
