+++ sxrenv
@@ -1,0 +1,6 @@
+#!/bin/bash
+# Source this to load the full environment that hutch python uses
+HERE=`dirname $(readlink -f $0)`
+source "${HERE}/sxrversion"
+pathmunge "${CONDA_BASE}/bin"
+source activate "${CONDA_ENVNAME}"
