+++ sxrpython
@@ -1,0 +1,6 @@
+#!/bin/bash
+# Launch hutch-python with all devices
+HERE=`dirname $(readlink -f $0)`
+source "${HERE}/sxrversion"
+pathmunge "${CONDA_BIN}"
+hutch-python --cfg "${HERE}/conf.yml" $@
