+++ sxrversion
@@ -1,0 +1,16 @@
+#!/bin/bash
+# Define the current conda environment.
+# This is referenced by the three launcher scripts.
+
+# edit this line only
+export CONDA_ENVNAME='pcds-0.6.1'
+
+export CONDA_BASE='/reg/g/pcds/pyps/conda/py36'
+export CONDA_ENV="${CONDA_BASE}/envs/${CONDA_ENVNAME}"
+export CONDA_BIN="${CONDA_ENV}/bin/wrappers/conda"
+
+HERE=`dirname $(readlink -f $0)`
+export PYTHONPATH="${HERE}:${HERE}/dev/devpath"
+export LD_LIBRARY_PATH=''
+
+source /reg/g/pcds/setup/pathmunge.sh
